chore: moves otel implementation into a separate file
